06_product_matching/requirements.txt
psycopg2-binary==2.9.9
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
tqdm==4.66.1
openai==1.3.0
numpy==1.24.3